Move BackupContext into useBackupContext.ts, remove types.ts
