--- src/ui/pages/Backup/types.ts
@@ -1,8 +1,0 @@
-export type BackupContext =
-  | {
-      appMode: 'onboarding';
-    }
-  | {
-      appMode: 'wallet';
-      groupId: string;
-    };
